refactor: remove Hello World folder
This commit removes the 'Hello World' folder from the project. This folder originally served as a basic introduction and starting point for the application. However, it is no longer necessary due to recent changes and expansions in our codebase. We've decided to remove this to avoid confusion and to maintain focus on the more advanced features of our app.